Добавляет компонент GameScreen, обновляет компоненты Artist и Genre
