+++ src/components/app/app.jsx
@@ -3,6 +3,7 @@ import {Switch, Route, BrowserRouter} from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 import ArtistQuestionScreen from '../artist-question-screen/artist-question-screen.jsx';
+import GameScreen from '../game-screen/game-screen.jsx';
 import GenreQuestionScreen from '../genre-question-screen/genre-question-screen.jsx';
 import WelcomeScreen from '../welcome-screen/welcome-screen.jsx';
 import {GameType, START_STEP} from '../../common/consts';
@@ -49,17 +50,25 @@ class App extends PureComponent {
       switch (question.type) {
         case GameType.ARTIST:
           return (
-            <ArtistQuestionScreen
-              question={question}
-              onAnswer={this._handleScreenChange}
-            />
+            <GameScreen
+              type={question.type}
+            >
+              <ArtistQuestionScreen
+                question={question}
+                onAnswer={this._handleScreenChange}
+              />
+            </GameScreen>
           );
         case GameType.GENRE:
           return (
-            <GenreQuestionScreen
-              question={question}
-              onAnswer={this._handleScreenChange}
-            />
+            <GameScreen
+              type={question.type}
+            >
+              <GenreQuestionScreen
+                question={question}
+                onAnswer={this._handleScreenChange}
+              />
+            </GameScreen>
           );
       }
     }
