+++ src/components/game-screen/game-screen.test.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import GameScreen from './game-screen.jsx';
+import {GameType} from '../../common/consts';
+
+const children = <div className="children-component" />;
+
+describe(`GameScreen component render correctly`, () => {
+  it(`with type GameType.ARTIST`, () => {
+    const tree = renderer.create(
+        <GameScreen
+          type={GameType.ARTIST}
+        >
+          {children}
+        </GameScreen>
+    ).toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+
+  it(`with type GameType.GENRE`, () => {
+    const tree = renderer.create(
+        <GameScreen
+          type={GameType.GENRE}
+        >
+          {children}
+        </GameScreen>
+    ).toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+});
