+++ src/components/genre-question-screen/genre-question-screen.jsx
@@ -3,12 +3,6 @@ import PropTypes from 'prop-types';
 
 import {GameType} from '../../common/consts';
 
-const timerStyle = {
-  filter: `url(#blur)`,
-  transform: `rotate(-90deg) scaleY(-1)`,
-  transformOrigin: `center`,
-};
-
 class GenreQuestionScreen extends PureComponent {
   constructor(props) {
     super(props);
@@ -73,31 +67,15 @@ class GenreQuestionScreen extends PureComponent {
     } = question;
 
     return (
-      <section className="game game--genre">
-        <header className="game__header">
-          <a className="game__back" href="#">
-            <span className="visually-hidden">Сыграть ещё раз</span>
-            <img className="game__logo" src="img/melody-logo-ginger.png" alt="Угадай мелодию"/>
-          </a>
-          <svg xmlns="http://www.w3.org/2000/svg" className="timer" viewBox="0 0 780 780">
-            <circle className="timer__line" cx="390" cy="390" r="370" style={timerStyle}/>
-          </svg>
-          <div className="game__mistakes">
-            <div className="wrong"/>
-            <div className="wrong"/>
-            <div className="wrong"/>
-          </div>
-        </header>
-        <section className="game__screen">
-          <h2 className="game__title">Выберите {genre} треки</h2>
-          <form
-            className="game__tracks"
-            onSubmit={this._handleSubmit(onAnswer, question)}
-          >
-            {this._renderTracks(answers, userAnswers)}
-            <button className="game__submit button" type="submit">Ответить</button>
-          </form>
-        </section>
+      <section className="game__screen">
+        <h2 className="game__title">Выберите {genre} треки</h2>
+        <form
+          className="game__tracks"
+          onSubmit={this._handleSubmit(onAnswer, question)}
+        >
+          {this._renderTracks(answers, userAnswers)}
+          <button className="game__submit button" type="submit">Ответить</button>
+        </form>
       </section>
     );
   }
